# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Onfly Travel API",
        "description": "API for managing travel orders and notifications",
        "contact": {
            "email": "support@onfly.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/me/notificacoes": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Get all notifications for authenticated user",
                "operationId": "0a52c5f88f64a6fbd469d32e63c61b1c",
                "responses": {
                    "200": {
                        "description": "List of notifications",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "550e8400-e29b-41d4-a716-446655440000"
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "example": "App\\\\Notifications\\\\TravelOrderStatusChanged"
                                                    },
                                                    "data": {
                                                        "properties": {
                                                            "travel_order_id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "destination": {
                                                                "type": "string",
                                                                "example": "São Paulo"
                                                            },
                                                            "old_status": {
                                                                "type": "string",
                                                                "example": "solicitado"
                                                            },
                                                            "new_status": {
                                                                "type": "string",
                                                                "example": "aprovado"
                                                            },
                                                            "message": {
                                                                "type": "string",
                                                                "example": "Seu pedido de viagem para São Paulo foi aprovado."
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "read_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "nullable": true
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/me/notificacoes/{id}/read": {
            "patch": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Mark notification as read",
                "operationId": "73f8eb86e2c1cc7f78e115659e395027",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification marked as read",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Notificação marcada como lida."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "550e8400-e29b-41d4-a716-446655440000"
                                                },
                                                "read_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Notification not found"
                    },
                    "403": {
                        "description": "Forbidden - not your notification"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your JWT token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}